\name{bhts2HTML}
\alias{bhts2HTML}

\title{
 Convert to HTML
}
\description{
  This function creates an HTML file.
}
\usage{
bhts2HTML(dat, dir, fname, title=NULL, bgcolor="#BBBBEE")
}
\arguments{
  \item{dat}{
  An object which is the output of \eqn{bhts()}.
}
  \item{dir}{
  Directory in which to store the file.
}
  \item{fname}{
  File name.
}
  \item{title}{
  The title of the html file.
}
  \item{bgcolor}{
  Color for the html background.
}
}

\examples{
 #See package vignette
}
