\name{BHTSpack-package}
\alias{BHTSpack-package}
\alias{BHTSpack}
\docType{package}
\title{
\packageTitle{BHTSpack}
}
\description{
Can be used for joint identification of candidate hits from multiple assays, in drug discovery. This package implements the framework of I. D. Shterev, D. B. Dunson, C. Chan and G. D. Sempowski. "Bayesian Multi Plate High Throughput Screening of Compounds", arXiv:1709.10041, September 2017. This project was funded by the Division of Allergy, Immunology, and Transplantation, National Institute of Allergy and Infectious Diseases, National Institutes of Health, Department of Health and Human Services, under contract No. HHSN272201400054C entitled "Adjuvant Discovery For Vaccines Against West Nile Virus and Influenza", awarded to Duke University and lead by Drs. Herman Staats and Soman Abraham.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BHTSpack}
\packageIndices{BHTSpack}

}
\author{
  I. D. Shterev, D. B. Dunson, C. Chan and G. D. Sempowski

Maintainer: I. D. Shterev <i.shterev@duke.edu>
}
\references{
   I. D. Shterev, D. B. Dunson, C. Chan and G. D. Sempowski. 
   "Bayesian Multi-Plate High-Throughput Screening of Compounds". \emph{Scientific Reports} to appear.
}

\keyword{ package }
