\name{nu.u}
\alias{nu.u}

\title{
 package internal function
}
\description{
 package internal function
}
\usage{
nu.u(ind, tau, H)
}

\arguments{
  \item{ind}{

}
  \item{tau}{

}
  \item{H}{

}
}

\value{
 Describe
}

\examples{
  H = 5
  n = 100
  tau = rgamma(1, 1, 1)
  ind = sample(H, n, replace=TRUE)
  nu.u(ind, tau, H)
}
