\name{pai.u}
\alias{pai.u}

\title{
 package internal function
}
\description{
 package internal function
}
\usage{
pai.u(b, a.pai, b.pai)
}

\arguments{
  \item{b}{

}
  \item{a.pai}{

}
  \item{b.pai}{

}
}

\value{
 Describe
}

\examples{
  n = 100
  b = rbinom(n, 1, 0.5)
  a.pai = 10^-6
  b.pai = 10^-6
  pai.u(b, a.pai, b.pai)
}
