\name{tau.u}
\alias{tau.u}

\title{
 package internal function
}
\description{
 package internal function
}
\usage{
tau.u(nu, a0, b0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{nu}{
  Description
}
  \item{a0}{
  Description
}
  \item{b0}{
  Description
}
}

\examples{
  K = 5
  a = 10^-6
  b = 10^-6
  nu = rbeta(K, a, b)
  tau.u(nu, a, b)
}
